[cmake] Add variables so that msbuild find the gsl
On Windows, the GNU Scientific Library gets installed in a specific
folder. So for msbuild.exe to find it, we add more variables that can be
used when invoking cmake. For example:

cmake -D -D GSL_INCLUDE_DIR="C:\Program Files (x86)\GnuWin32\include" ...

This is needed because of error message `gsl_errno.h: no such file or
directory`