sinktaskmgr: prevent to send hole in history result.
Due to the multithread tasks, result may be set in the list in different
order.
i.e result of frame 5 may be set before frame 3.
This modification stop coping the result in case of a hole in
result. i.e the result's frame number will be lower than the request.
This will only work if the client follow the real time.